moving hooks for reported runtime console logs and runtimeErrors into a dedicated file
